update ddk includes
git-svn-id: svn://kolibrios.org@2967 a494cfbc-eb01-0410-851d-a64ba20cac60

+++ drivers/include/linux/types.h
@@ -246,7 +246,7 @@ struct ustat {
 
 typedef unsigned char        u8_t;
 typedef unsigned short       u16_t;
-typedef unsigned int         u32_t;
+typedef unsigned long        u32_t;
 typedef unsigned long long   u64_t;
 
 typedef unsigned int         addr_t;
